- Cambios estructurales

+++ src/main/java/quiniela/model/TeamGroup.java
@@ -3,8 +3,8 @@ package quiniela.model;
 public class TeamGroup {
 
     private int points = 0;
-    private int positiveGoals = 0;
-    private int negativeGoals = 0;
+    private long positiveGoals = 0;
+    private long negativeGoals = 0;
 
     public int getPoints() {
         return points;
@@ -14,19 +14,19 @@ public class TeamGroup {
         this.points = points;
     }
 
-    public int getPositiveGoals() {
+    public long getPositiveGoals() {
         return positiveGoals;
     }
 
-    public void setPositiveGoals(int positiveGoals) {
+    public void setPositiveGoals(long positiveGoals) {
         this.positiveGoals = positiveGoals;
     }
 
-    public int getNegativeGoals() {
+    public long getNegativeGoals() {
         return negativeGoals;
     }
 
-    public void setNegativeGoals(int negativeGoals) {
+    public void setNegativeGoals(long negativeGoals) {
         this.negativeGoals = negativeGoals;
     }
 }
